Time: 55 ms (33.96%), Space: 10.3 MB (33.96%) - LeetHub

+++ 3751-total-waviness-of-numbers-in-range-i/3751-total-waviness-of-numbers-in-range-i.cpp
@@ -4,7 +4,7 @@ public:
         string s = to_string(num);
         int cnt = 0;
 
-        for (int i = 1; i < (int)s.size() - 1; i++) {
+        for (int i = 1; i <s.size() - 1; i++) {
             if ((s[i] > s[i - 1] && s[i] > s[i + 1]) ||
                 (s[i] < s[i - 1] && s[i] < s[i + 1])) {
                 cnt++;
